Important Bug in Open Method Fixed
Whenever the open method threw an exception (more exact: after the factory's createInstance method returned normally), a cleanup was not possible: Service could not be closed (because it has not been opened successfully before), but further open failed as well (because a broken version of the service has been created previously).
This has been fixed: If open fails after createInstance ran successfully, a cleanup will be performed internally.