Enhancement Plausibility Checks
For read-write properties that can be accessed after open, UPOS does not specify how write access shall be handled in exclusive-use devices when the device has not been claimed.
As default, the service checks only if the device has been opened. It is completely up to the service to decide whether trying to change such a property while the device is not claimed results in an exception or in a property change.
Now, if you set AllowAlwaysSetProperties to false in jpos.xml, the property setter will only be called if the new value is equal to the previous value of the device has been claimed. Otherwise, an exception will be called.
However, there are some exceptions. If the property specifies only some change in the interface between application and the service, only the open check will be made. This is the case for properties like for example MapCharacterSet or FlagWhenIdle.