Support For UPOS 1.16
Support for UPOS 1.16 added: Until no other  jpos controls for JavaPOS 1.16 is available, module Jpos116Dummy can be used to provide JavaPOS 1.16 functionality to device classes Lights and POSPower. Because this implementation uses the javapos 1.15.1, A jar file created from that module must be placed before the jpos 1.15 jar file to replace the corresponding controls and constants. When a full jpos 1.16 implementatiomn becomes available, this module (and the corresponding jar file) can be removed.