Enhancements FiscalPrinter
New methods added to FiscalPrinter to allow the service to retrieve and to pass numerical values as int or long instead of String. This is important for Currency values where the interpretation as String is not clear.
Two wrapper classes have been added which allows on-the-fly conversion of currency values in String parameters for application and service implementations which use different interpretations of Currency values in strings.