Changes due to Specification Ambiguities
Due to different opinions between me and UPOS members, a new property NoPositionRangeAvailabilityCheck for device class GestureControl has been added to jpos.xml. This allows to implement services which handle setPosition and getPosition method calls with jointID values of jounts that do not have a position range as an error or to allow such jointIDs. However, how position values of such joints will be handled if NoPositionRangeAvailabilityCheck has been set to true in jpos.xml, depends on the device specific service implementation.